Fix timeout assignment for implicit settle

+++ WifiAutologin/WebDriver.cs
@@ -71,7 +71,7 @@ public class WebDriver : IDisposable
         Logger.Debug("Allowing page a few seconds to settle after login...");
         ActOnPage(new Config.NetworkAction {
           Action = Config.NetworkActionType.Settle,
-          Timeout = 2.0
+          Timeout = (float)2.0
         });
     }
